requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.1
Django==4.0
django-crispy-forms==1.14.0
flake8==5.0.4
gunicorn==20.1.0
idna==3.3
install==1.3.5
isort==5.10.1
mccabe==0.7.0
Pillow==9.2.0
pycodestyle==2.9.1
pyflakes==2.5.0
requests==2.28.1
sorl-thumbnail==12.9.0
sqlparse==0.4.2
urllib3==1.26.12
psycopg2-binary==2.8.6